Fix TreeMultiset.subMultiset().size() overflow condition. From Louis: http://codereview.appspot.com/5153041/
Also, remove SortedTreeMultiset now that it's been moved to TreeMultiset, since it, too, would have needed to have been updated in this CL.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=24372922

+++ guava-tests/test/com/google/common/collect/BstRangeOpsTest.java
@@ -60,7 +60,7 @@ public class BstRangeOpsTest extends TestCase {
   public void testCountInRangeLowerBound() {
     for (char c : "abcdefg".toCharArray()) {
       for (BoundType type : BoundType.values()) {
-        int count = BstRangeOps.totalInRange(
+        long count = BstRangeOps.totalInRange(
             countAggregate, GeneralRange.downTo(Ordering.natural(), c, type), ROOT);
         char d = c;
         if (type == BoundType.OPEN) {
@@ -74,7 +74,7 @@ public class BstRangeOpsTest extends TestCase {
   public void testCountInRangeUpperBound() {
     for (char c : "abcdefg".toCharArray()) {
       for (BoundType type : BoundType.values()) {
-        int count = BstRangeOps.totalInRange(
+        long count = BstRangeOps.totalInRange(
             countAggregate, GeneralRange.upTo(Ordering.natural(), c, type), ROOT);
         char d = c;
         if (type == BoundType.CLOSED) {
@@ -94,7 +94,7 @@ public class BstRangeOpsTest extends TestCase {
             if (i == j && lb == BoundType.OPEN && ub == BoundType.OPEN) {
               continue;
             }
-            int count = BstRangeOps.totalInRange(countAggregate, GeneralRange.range(
+            long count = BstRangeOps.totalInRange(countAggregate, GeneralRange.range(
                 Ordering.natural(), chars.charAt(i), lb, chars.charAt(j), ub), ROOT);
             char lo = chars.charAt(i);
             if (lb == BoundType.OPEN) {
